use json encoder
